backend/database/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
psycopg2-binary==2.9.11
pydantic==2.12.5
gunicorn==23.0.0
